chore: Move /me route to user module
