+++ api.planx.uk/modules/misc/controller.ts
@@ -1,38 +1,7 @@
 import { RequestHandler } from "express";
-import { getClient } from "../../client";
-import { userContext } from "../auth/middleware";
-import { ServerError } from "../../errors";
 import { stringify } from "csv-stringify";
 import { z } from "zod";
 import { ValidatedRequestHandler } from "../../shared/middleware/validate";
-
-export const getLoggedInUserDetails: RequestHandler = async (
-  _req,
-  res,
-  next,
-) => {
-  try {
-    const $client = getClient();
-
-    const id = userContext.getStore()?.user.sub;
-    if (!id)
-      throw new ServerError({
-        message: "User ID missing from request",
-        status: 400,
-      });
-
-    const user = await $client.user.getById(parseInt(id));
-    if (!user)
-      throw new ServerError({
-        message: `Unable to locate user with ID ${id}`,
-        status: 400,
-      });
-
-    res.json(user);
-  } catch (error) {
-    next(error);
-  }
-};
 
 export const healthCheck: RequestHandler = (_req, res) =>
   res.json({ hello: "world" });
