+++ api.planx.uk/modules/misc/docs.yaml
@@ -24,94 +24,39 @@ paths:
                     type: string
                 example:
                   hello: world
-  /me:
-    get:
-      summary: Get information about currently logged in user
-      tags:
-        - misc
-      security:
-        - bearerAuth: []
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  id:
-                    type: integer
-                    format: int32
-                    example: 123
-                  firstName:
-                    type: string
-                    example: Albert
-                  lastName:
-                    type: string
-                    example: Einstein
-                  email:
-                    type: string
-                    example: albert@princeton.edu
-                  isPlatformAdmin:
-                    type: boolean
-                    example: true
-                  teams:
-                    type: array
-                    items:
-                      type: object
-                      properties:
-                        team:
-                          type: object
-                          properties:
-                            id:
-                              type: integer
-                              format: int32
-                              example: 123
-                            slug:
-                              type: string
-                              example: opensystemslab
-                            name:
-                              type: string
-                              example: Open Systems Lab
-                        role:
-                          type: string
-                          enum: ["teamEditor", "teamViewer"]
-                          example: "teamEditor"
-        "401":
-          $ref: "#/components/responses/Unauthorised"
-  /download-application:
-    post:
-      summary: Download application
-      description: Allows an applicant to download their application data on the Confirmation page
-      tags:
-        - misc
-      requestBody:
-        required: true
+/download-application:
+  post:
+    summary: Download application
+    description: Allows an applicant to download their application data on the Confirmation page
+    tags:
+      - misc
+    requestBody:
+      required: true
+      content:
+        application/json:
+          schema:
+            type: object
+            properties:
+              data:
+                required: true
+                type: array
+                items:
+                  type: object
+                  properties:
+                    question:
+                      type: string
+                      required: true
+                    responses:
+                      type: any
+                      required: true
+                    metadata:
+                      type: any
+                      required: false
+    responses:
+      "200":
         content:
-          application/json:
+          text/csv:
             schema:
-              type: object
-              properties:
-                data:
-                  required: true
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      question:
-                        type: string
-                        required: true
-                      responses:
-                        type: any
-                        required: true
-                      metadata:
-                        type: any
-                        required: false
-      responses:
-        "200":
-          content:
-            text/csv:
-              schema:
-                type: string
-        "500":
-          $ref: "#/components/responses/ErrorMessage"
+              type: string
+      "500":
+        $ref: "#/components/responses/ErrorMessage"
