+++ api.planx.uk/modules/user/index.test.ts
@@ -1,6 +1,9 @@
 import supertest from "supertest";
 import app from "../../server";
-import { authHeader } from "../../tests/mockJWT";
+import { authHeader, getJWT } from "../../tests/mockJWT";
+import { userContext } from "../auth/middleware";
+
+const getStoreMock = jest.spyOn(userContext, "getStore");
 
 const mockUser = {
   firstName: "Bilbo",
@@ -12,6 +15,23 @@ const mockUser = {
 const mockCreateUser = jest.fn();
 const mockDeleteUser = jest.fn();
 const mockGetByEmail = jest.fn().mockResolvedValue(mockUser);
+const mockGetById = jest.fn().mockResolvedValue({
+  id: 123,
+  firstName: "Albert",
+  lastName: "Einstein",
+  email: "albert@princeton.edu",
+  isPlatformAdmin: true,
+  teams: [
+    {
+      teamId: 1,
+      role: "teamEditor",
+    },
+    {
+      teamId: 24,
+      role: "teamEditor",
+    },
+  ],
+});
 
 jest.mock("@opensystemslab/planx-core", () => {
   return {
@@ -20,6 +40,7 @@ jest.mock("@opensystemslab/planx-core", () => {
         create: () => mockCreateUser(),
         delete: () => mockDeleteUser(),
         getByEmail: () => mockGetByEmail(),
+        getById: () => mockGetById(),
       },
     })),
   };
@@ -124,6 +145,72 @@ describe("Delete user endpoint", () => {
         expect(mockDeleteUser).toHaveBeenCalled();
         expect(res.body).toHaveProperty("message");
         expect(res.body.message).toMatch(/Successfully deleted user/);
+      });
+  });
+});
+
+describe("/me endpoint", () => {
+  beforeEach(() => {
+    getStoreMock.mockReturnValue({
+      user: {
+        sub: "123",
+        jwt: getJWT({ role: "teamEditor" }),
+      },
+    });
+  });
+
+  it("returns an error if authorization headers are not set", async () => {
+    await supertest(app)
+      .get("/user/me")
+      .expect(401)
+      .then((res) => {
+        expect(res.body).toEqual({
+          error: "No authorization token was found",
+        });
+      });
+  });
+
+  it("returns an error for invalid user context", async () => {
+    getStoreMock.mockReturnValue({
+      user: {
+        sub: undefined,
+        jwt: getJWT({ role: "teamEditor" }),
+      },
+    });
+
+    await supertest(app)
+      .get("/user/me")
+      .set(authHeader({ role: "teamEditor" }))
+      .expect(400)
+      .then((res) => {
+        expect(res.body).toEqual({
+          error: "User ID missing from request",
+        });
+      });
+  });
+
+  it("returns an error for an invalid email address", async () => {
+    mockGetById.mockResolvedValueOnce(null);
+
+    await supertest(app)
+      .get("/user/me")
+      .set(authHeader({ role: "teamEditor" }))
+      .expect(400)
+      .then((res) => {
+        expect(res.body).toEqual({
+          error: "Unable to locate user with ID 123",
+        });
+      });
+  });
+
+  it("returns user details for a logged in user", async () => {
+    await supertest(app)
+      .get("/user/me")
+      .set(authHeader({ role: "teamEditor" }))
+      .expect(200)
+      .then((res) => {
+        expect(res.body).toHaveProperty("email", "albert@princeton.edu");
+        expect(res.body.teams).toHaveLength(2);
       });
   });
 });
