+++ api.planx.uk/modules/user/routes.ts
@@ -1,17 +1,28 @@
 import { Router } from "express";
-import { usePlatformAdminAuth } from "../auth/middleware";
+import { useLoginAuth, usePlatformAdminAuth } from "../auth/middleware";
 import { validate } from "../../shared/middleware/validate";
 import {
   createUserSchema,
   createUser,
   deleteUserSchema,
   deleteUser,
+  getLoggedInUserDetails,
 } from "./controller";
 
 const router = Router();
 
-router.use("/user", usePlatformAdminAuth);
-router.put("/user", validate(createUserSchema), createUser);
-router.delete("/user/:email", validate(deleteUserSchema), deleteUser);
+router.put(
+  "/user",
+  usePlatformAdminAuth,
+  validate(createUserSchema),
+  createUser,
+);
+router.delete(
+  "/user/:email",
+  usePlatformAdminAuth,
+  validate(deleteUserSchema),
+  deleteUser,
+);
+router.get("/user/me", useLoginAuth, getLoggedInUserDetails);
 
 export default router;
